backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
yt-dlp
opencv-python
pytesseract
Pillow
reportlab
python-docx
openai
python-dotenv
pydantic
aiofiles
numpy
torch
torchaudio
ffmpeg-python
